WIP: refactor to handle subscription method
