+++ pubsub/src/delegates.rs
@@ -39,77 +39,6 @@ impl<M, T, F, I> jsonrpc_pubsub::UnsubscribeRpcMethod<M> for DelegateUnsubscribe
 	}
 }
 
-/// A set of RPC methods and notifications tied to single `delegate` struct.
-pub struct IoDelegate<T, M = ()> where
-	T: Send + Sync + 'static,
-	M: Metadata,
-{
-	delegate: Arc<T>,
-	methods: HashMap<String, RemoteProcedure<M>>,
-}
-
-impl<T, M> IoDelegate<T, M> where
-	T: Send + Sync + 'static,
-	M: Metadata,
-{
-	/// Creates new `IoDelegate`
-	pub fn new(delegate: Arc<T>) -> Self {
-		IoDelegate {
-			delegate: delegate,
-			methods: HashMap::new(),
-		}
-	}
-
-	/// Adds an alias to existing method.
-	/// NOTE: Aliases are not transitive, i.e. you cannot create alias to an alias.
-	pub fn add_alias(&mut self, from: &str, to: &str) {
-		self.methods.insert(from.into(), RemoteProcedure::Alias(to.into()));
-	}
-
-	/// Adds async method to the delegate.
-	pub fn add_method<F, I>(&mut self, name: &str, method: F) where
-		F: Fn(&T, Params) -> I,
-		I: IntoFuture<Item = Value, Error = Error>,
-		F: Send + Sync + 'static,
-		I::Future: Send + 'static,
-	{
-		self.methods.insert(name.into(), RemoteProcedure::Method(Arc::new(
-			DelegateAsyncMethod {
-				delegate: self.delegate.clone(),
-				closure: method,
-			}
-		)));
-	}
-
-	/// Adds async method with metadata to the delegate.
-	pub fn add_method_with_meta<F, I>(&mut self, name: &str, method: F) where
-		F: Fn(&T, Params, M) -> I,
-		I: IntoFuture<Item = Value, Error = Error>,
-		F: Send + Sync + 'static,
-		I::Future: Send + 'static,
-	{
-		self.methods.insert(name.into(), RemoteProcedure::Method(Arc::new(
-			DelegateMethodWithMeta {
-				delegate: self.delegate.clone(),
-				closure: method,
-			}
-		)));
-	}
-
-	/// Adds notification to the delegate.
-	pub fn add_notification<F>(&mut self, name: &str, notification: F) where
-		F: Fn(&T, Params),
-		F: Send + Sync + 'static,
-	{
-		self.methods.insert(name.into(), RemoteProcedure::Notification(Arc::new(
-			DelegateNotification {
-				delegate: self.delegate.clone(),
-				closure: notification,
-			}
-		)));
-	}
-}
-
 impl<T, M> IoDelegate<T, M> where
 	T: Send + Sync + 'static,
 	M: PubSubMetadata,
